backend/requirements.txt
# ===============================
# Core FastAPI stack
# ===============================
fastapi==0.110.0
uvicorn==0.29.0

# ===============================
# Database & ORM
# ===============================
sqlalchemy==2.0.29
alembic==1.13.1

# Choose ONE database driver
psycopg2-binary==2.9.9      # PostgreSQL (recommended)
# mysqlclient==2.2.4       # MySQL (use instead if needed)

# ===============================
# Data validation & settings
# ===============================
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ===============================
# Authentication & Security
# ===============================
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ===============================
# Utilities
# ===============================
email-validator==2.1.1
